DiscordBot/requirements.txt
discord.py
aiomysql